requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
Pillow
drf-yasg
django-imagekit
PyYAML
